Initial check for invalid JNI method signatures
Summary:
When using QAndroidJniObject to call Android API one needs to specify method signatures using strings. This is error-prone and errors are usually only detected at runtime.
This check searches for a set of known methods that take such signatures and validates them using a regular expression. If a signature is invalid a warning is emitted.

Test Plan: Ran tests

Differential Revision: https://phabricator.kde.org/D28887